Transfer my old GLU project from 2011 to GIT
The personalization has supposedy been removed. Publishing my own code with MIT licence, so it can be used, but I won't be abused. It's a glucode diary for diabetics, which at the moment I'm not. So NO RESPONSIBLITY will be assumed, likewise pen, paper manufacturers nor form printers will take no shit for use of the forms whatsoever.